drop pointless include file in hopes it might fix clang-tidy.
